+++ ports/silabs/common-hal/analogio/AnalogIn.c
@@ -184,7 +184,7 @@ void IADC_IRQHandler(void) {
 //  Get adc value, use IADC_IRQHandler
 //  adc value 0 - 65535
 uint16_t common_hal_analogio_analogin_get_value(analogio_analogin_obj_t *self) {
-    uint16_t avg_value = 0;
+    uint32_t avg_value = 0;
     for (int i = 0; i < self->sample_size; i++) {
         // Start scan
         IADC_command(IADC0, iadcCmdStartScan);
